+++ packages/EasyErrorHandler/src/Bridge/Symfony/DependencyInjection/Compiler/ApiPlatformCompilerPass.php
@@ -17,7 +17,10 @@ final class ApiPlatformCompilerPass implements CompilerPassInterface
 
         $container->removeDefinition('api_platform.listener.exception.validation');
 
+        // Set `handle_all_throwables` to `true`
+        // We need this to handle \TypeError in \EonX\EasyErrorHandler\Bridge\Symfony\Builder\ApiPlatformValidationErrorResponseBuilder
         // See \Symfony\Component\HttpKernel\HttpKernel
+        // See https://symfony.com/doc/current/reference/configuration/framework.html#handle-all-throwables
         $container->getDefinition('http_kernel')
             ->replaceArgument(4, true);
     }
