+++ packages/EasyErrorHandler/tests/DataProviders/TestRenderWithDefaultBuildersDataProvider.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace EonX\EasyErrorHandler\Tests;
+namespace EonX\EasyErrorHandler\Tests\DataProviders;
 
 use EonX\EasyErrorHandler\Tests\Stubs\BaseExceptionStub;
 use EonX\EasyErrorHandler\Tests\Stubs\ValidationExceptionStub;
